requirements.txt
pyzmq>=25.0
numpy>=1.24
requests>=2.28
